Fixed installation bugs

+++ install/system.sh
@@ -8,7 +8,7 @@ if [[ "${suser}" != "root" ]]; then
   exit 1
 fi
 
-sudo pacman -S --needed base-devel yajl
+sudo pacman -S --needed base-devel yajl wget
 
 if [[ -z "$(pacman -Qs yaourt)" ]]; then
   echo "Installing Yaourt"
@@ -16,19 +16,21 @@ if [[ -z "$(pacman -Qs yaourt)" ]]; then
   mkdir -p /tmp/yaourt-install
   cd /tmp/yaourt-install
 
-  wget https://aur.archlinux.org/cgit/aur.git/snapshot/package-query.tar.gz
-  tar zxvf package-query.tar.gz
-  cd package-query
-  makepkg
-  sudo pacman -U package-query*.tar.xz
-  cd ..
-  rm -r package-query
+  if [[ -z "$(pacman -Qs package-query)" ]]; then
+    wget https://aur.archlinux.org/cgit/aur.git/snapshot/package-query.tar.gz
+    tar zxvf package-query.tar.gz
+    cd package-query
+    makepkg
+    sudo pacman -U package-query*.tar.xz
+    cd ..
+    rm -r package-query
+  fi
 
   wget https://aur.archlinux.org/cgit/aur.git/snapshot/yaourt.tar.gz
   tar zxvf yaourt.tar.gz
   cd yaourt
   makepkg
-  sudo pacman -U yaourt*.tar.xz
+  sudo pacman -U yaourt*.pkg.tar.xz
   cd ..
   rm -r yaourt
